Add NetworkEndian alias
Network byte order is defined by [RFC 1700][1] to be big-endian, and is
referred to in several protocol specifications.  This type is an alias
of BigEndian.  This alias can be used to make it more clear why you're
using a particular byte order in code, if you are implementing a
specification that refers to network byte order.

[1]: https://tools.ietf.org/html/rfc1700