Fix prop_ext_[u]int_*::native_endian on BE targets
The similar `big_endian` tests were using an offset to read from the
end of the written `u64`, but the `native_endian` tests were reading
directly, just like the `little_endian` tests.  That's of course only
correct when the target actually is little endian.

That `big_endian` offset is now sliced directly, instead of cloning into
another vector, and then this logic is also used in the `native_endian`
test, depending on the current `#[cfg(target_endian)]`.

Fixes #102.