i128: enable support for 128-bit integers automatically
This adds a build.rs to byteorder that will set a conditional compilation
flag automatically if the current Rust compiler supports 128-bit integers.

This makes the i128 feature itself a no-op. We continue to allow the
feature to be supplied for backwards compatibility.

Addresses https://github.com/TyOverby/bincode/issues/250