msrv: bump to 1.41.1
CI was failing on Rust 1.12 because it looks like it can no longer
update the registry. I see no reason to continue maintaining support
for such an old version of Rust.

So the only problem is to pick the new MSRV. Rust 1.41.1 was release
about a year ago and is also the version of Rust packaged in Debian
Buster.

This commit tries to stay to cosmetic changes. e.g., Changing `try!`
to `?`, removing version sniffing and so on. A subsequent commit will
start adapting the code to use newer APIs from std.