impl: prefix unsafe macros with 'unsafe_'
If these could be defined as functions, then we would have considered
them to be unsound. Since there is no such thing as an "unsafe macro" in
Rust, we adopt a convention that such a macro start with 'unsafe_'.

There are no changes in the safety properties of the implementation in
this commit. This is only about making the code a bit clearer and more
consistent with Rust's safety story.

Kudos to @tspiteri for the idea:
https://internals.rust-lang.org/t/hidden-unsafe-due-to-unintentionally-abusable-macros-and-include/14107/26